move finality transition related tests from api_tests.cpp to finality_transition_tests.cpp
